Added script for checking the types.
