+++ src/package.json
@@ -108,7 +108,8 @@
     "test": "mocha --import=tsx --timeout 120000 --recursive tests/backend/specs ../node_modules/ep_*/static/tests/backend/specs",
     "test-container": "mocha --import=tsx --timeout 5000 tests/container/specs/api",
     "dev": "node --import tsx node/server.ts",
-    "prod": "node --import tsx node/server.ts"
+    "prod": "node --import tsx node/server.ts",
+    "ts-check": "tsc --noEmit --watch"
   },
   "version": "1.9.6",
   "license": "Apache-2.0"
